requirements.txt
esp32-machine-emulator
